requirements.txt
numpy>=1.21.0
qiskit>=0.33.0  # Quantum computing framework
pennylane>=0.20.0  # Quantum ML framework
torch>=1.9.0  # Deep learning
tensorflow>=2.6.0  # Neural network processing
scipy>=1.7.0  # Scientific computing
pandas>=1.3.0  # Data manipulation
matplotlib>=3.4.0  # Visualization
seaborn>=0.11.0  # Statistical visualization
pytest>=6.2.5  # Testing
black>=21.7b0  # Code formatting
isort>=5.9.3  # Import sorting
flake8>=3.9.0  # Code linting
mypy>=0.910  # Type checking
asyncio>=3.4.3  # Asynchronous I/O
aiohttp>=3.8.0  # Async HTTP
pyyaml>=5.4.1  # YAML parsing
tqdm>=4.62.0  # Progress bars
scikit-learn>=0.24.0  # Machine learning
pyserial>=3.5  # Serial communication for hardware
